+++ backend/main.py
@@ -1,6 +1,23 @@
 from fastapi import FastAPI
+from app.modules.telemetry.mqtt_service import start_mqtt
+from contextlib import asynccontextmanager
 
-app = FastAPI()
+
+# Arrancamos MQTT al iniciar la App
+@asynccontextmanager
+async def lifespan(app: FastAPI):
+    # --- Lo que ocurre al ARRANCAR ---
+    print("Iniciando conexión MQTT...")
+    start_mqtt() 
+    
+    yield # Aquí es donde la aplicación "vive"
+    
+    # --- Lo que ocurre al CERRAR ---
+    print("Cerrando conexiones...")
+    # Aquí se puede poner client.disconnect() si quisiéramos ser súper limpios
+
+app = FastAPI(lifespan=lifespan)
+
 
 @app.get("/")
 async def read_root():